do not create a python package
... where there is no package.
Only a script was installed from the setup.py
but catkin provides a separate command for that
http://docs.ros.org/en/jade/api/catkin/html/howto/format2/installing_python.html

I retained the installation to the global path. It is unusual to do
that, but it has been like that for a long time.

This started failing with recent setuptools with the following error
message:

---
error: Multiple top-level packages discovered in a flat-layout: ['debian', 'config', 'launch'].

To avoid accidental inclusion of unwanted files or directories,
setuptools will not proceed with this build.

If you are trying to create a single distribution with multiple packages
on purpose, you should not rely on automatic discovery.
Instead, consider the following options:

1. set up custom discovery (`find` directive with `include` or `exclude`)
2. use a `src-layout`
3. explicitly set `py_modules` or `packages` with a list of names

To find more information, look for "package discovery" on setuptools docs.
---